ml-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
tensorflow>=2.18.0
Pillow>=10.2.0
numpy>=1.26.0
hypothesis>=6.92.1
pytest>=7.4.4
httpx>=0.26.0
